requirements.txt
boto3==1.34.128
pyyaml==6.0.1
